Refactoring and getting tests to pass
