+++ openprescribing/frontend/management/commands/measure_definitions/ktt12_diabetes_blood_glucose.json
@@ -18,14 +18,14 @@
   "is_cost_based": false, 
   "low_is_good": false, 
   "numerator_columns": [
-    "SUM(total_items) AS numerator, "
+    "SUM(items) AS numerator, "
   ], 
   "numerator_from": "[ebmdatalab:hscic.prescribing] ", 
   "numerator_where": [
     "((bnf_code LIKE '0601021%%')  OR (bnf_code LIKE '0601022%%'))  "
   ], 
   "denominator_columns": [
-    "SUM(total_items) AS denominator, "
+    "SUM(items) AS denominator, "
   ], 
   "denominator_from": "[ebmdatalab:hscic.prescribing] ", 
   "denominator_where": [
