+++ openprescribing/frontend/management/commands/measure_definitions/ktt9_cephalosporins.json
@@ -27,14 +27,14 @@
   "is_cost_based": false, 
   "low_is_good": true, 
   "numerator_columns": [
-    "SUM(total_items) AS numerator, "
+    "SUM(items) AS numerator, "
   ], 
   "numerator_from": "[ebmdatalab:hscic.prescribing] ", 
   "numerator_where": [
     "((bnf_code LIKE '0501013K0%%')  OR (bnf_code LIKE '0501021%%')  OR (bnf_code LIKE '050112%%'))  "
   ], 
   "denominator_columns": [
-    "SUM(total_items) AS denominator, "
+    "SUM(items) AS denominator, "
   ], 
   "denominator_from": "[ebmdatalab:hscic.prescribing] ", 
   "denominator_where": [
