+++ openprescribing/frontend/management/commands/measure_sql/practice_ratios.sql
@@ -60,7 +60,7 @@ FROM (
         (num.practice=denom.practice
           AND num.month=denom.month)) ratios
       INNER JOIN
-        ebmdatalab.hscic.practices
+        {practices_from} AS practices
       -- only examine standard GP practices
       ON practices.setting = 4 AND practices.code=ratios.practice_id AND (practices.open_date < FORMAT_TIMESTAMP('%Y-%m-%d', ratios.month) OR practices.open_date IS NULL) AND (practices.close_date > FORMAT_TIMESTAMP('%Y-%m-%d', ratios.month) or practices.close_date IS NULL)
       )
